Extract adjacent-identifiers rule into its own file
`preParse` now just walks the token stream and dispatches by token
type. The adjacent-identifiers rule (and its supporting helpers —
boundary set, suggestion logic, Levenshtein) lives in
`preParseAdjacentIdentifiers.ts`. Future rules follow the same shape:
one file per rule, called from `preParse` when its trigger token is
seen.
